Huge Shader work
Changed every shader used in each scene and added an outline to the
standard shader through a third party script.

Also added the toon_ prefix to each shader used in the scenes that use
the standard (outline) so we can easily search for them in the project
files and edit the outline color and thickness all together.